test : domain 레이어 테스트 소스 분리  - 클래스별 테스트케이스 분리
